zta/microservice6/requirements.txt
uvicorn
fastapi
pydantic
asyncio
aiohttp
pycryptodome